# content/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "EndingsEvaluateRequest": {
        "additionalProperties": false,
        "properties": {
          "at_absolute": {
            "minimum": 0.0,
            "title": "At Absolute",
            "type": "integer"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id",
          "at_absolute"
        ],
        "title": "EndingsEvaluateRequest",
        "type": "object"
      },
      "EndingsResponse": {
        "additionalProperties": false,
        "properties": {
          "endings": {
            "items": {
              "$ref": "#/components/schemas/EndingView"
            },
            "title": "Endings",
            "type": "array"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id"
        ],
        "title": "EndingsResponse",
        "type": "object"
      },
      "EraComposeRequest": {
        "additionalProperties": false,
        "properties": {
          "end_absolute": {
            "minimum": 1.0,
            "title": "End Absolute",
            "type": "integer"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "start_absolute": {
            "minimum": 0.0,
            "title": "Start Absolute",
            "type": "integer"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id",
          "owner_id",
          "start_absolute",
          "end_absolute"
        ],
        "title": "EraComposeRequest",
        "type": "object"
      },
      "EraView": {
        "additionalProperties": false,
        "properties": {
          "end_absolute": {
            "title": "End Absolute",
            "type": "integer"
          },
          "era_id": {
            "format": "uuid",
            "title": "Era Id",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "start_absolute": {
            "title": "Start Absolute",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "era_id",
          "owner_id",
          "start_absolute",
          "end_absolute",
          "text",
          "version"
        ],
        "title": "EraView",
        "type": "object"
      },
      "ErasResponse": {
        "additionalProperties": false,
        "properties": {
          "eras": {
            "items": {
              "$ref": "#/components/schemas/EraView"
            },
            "title": "Eras",
            "type": "array"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id"
        ],
        "title": "ErasResponse",
        "type": "object"
      },
...
      "FocusAssignRequest": {
        "additionalProperties": false,
        "properties": {
          "effective_absolute": {
            "minimum": 0.0,
            "title": "Effective Absolute",
            "type": "integer"
          },
          "from_character_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Character Id"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "slot": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Slot",
            "type": "string"
          },
          "to_character_id": {
            "format": "uuid",
            "title": "To Character Id",
            "type": "string"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id",
          "slot",
          "to_character_id",
          "reason",
          "effective_absolute"
        ],
        "title": "FocusAssignRequest",
        "type": "object"
      },
      "FocusAssignmentView": {
        "additionalProperties": false,
        "properties": {
          "effective_absolute": {
            "title": "Effective Absolute",
            "type": "integer"
          },
          "from_character_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Character Id"
          },
          "from_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Name"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "slot": {
            "title": "Slot",
            "type": "string"
          },
          "to_character_id": {
            "format": "uuid",
            "title": "To Character Id",
            "type": "string"
          },
          "to_name": {
            "title": "To Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "slot",
          "version",
          "to_character_id",
          "to_name",
          "effective_absolute",
          "reason"
        ],
        "title": "FocusAssignmentView",
        "type": "object"
      },
      "FocusResponse": {
        "additionalProperties": false,
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/FocusAssignmentView"
            },
            "title": "Assignments",
            "type": "array"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id"
        ],
        "title": "FocusResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MacroAdvanceRequest": {
        "additionalProperties": false,
        "properties": {
          "day": {
            "minimum": 1.0,
            "title": "Day",
            "type": "integer"
          },
          "resolution": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Resolution",
            "type": "string"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id",
          "day",
          "resolution"
        ],
        "title": "MacroAdvanceRequest",
        "type": "object"
      },
      "MacroAdvanceResponse": {
        "additionalProperties": false,
        "properties": {
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "end_absolute": {
            "title": "End Absolute",
            "type": "integer"
          },
          "event_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Event Ids",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "start_absolute": {
            "title": "Start Absolute",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "state",
          "start_absolute",
          "end_absolute",
          "duplicate"
        ],
        "title": "MacroAdvanceResponse",
        "type": "object"
      },
...
      "MacroRunsResponse": {
        "additionalProperties": false,
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/MacroRunView"
            },
            "title": "Runs",
            "type": "array"
          },
          "world_id": {
            "format": "uuid",
            "title": "World Id",
            "type": "string"
          }
        },
        "required": [
          "world_id"
        ],
        "title": "MacroRunsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "worldsim",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/macro/advance": {
      "post": {
        "operationId": "macro_advance_api_v1_macro_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MacroAdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MacroAdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Macro Advance",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/endings": {
      "get": {
        "operationId": "endings_api_v1_macro_endings_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Endings",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/endings/evaluate": {
      "post": {
        "operationId": "evaluate_endings_view_api_v1_macro_endings_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndingsEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Endings View",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/eras": {
      "get": {
        "operationId": "eras_api_v1_macro_eras_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_absolute",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Start Absolute",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end_absolute",
            "required": false,
            "schema": {
              "default": 3600,
              "minimum": 1,
              "title": "End Absolute",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErasResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Eras",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/eras/compose": {
      "post": {
        "operationId": "compose_era_view_api_v1_macro_eras_compose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EraComposeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EraView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compose Era View",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/focus": {
      "get": {
        "operationId": "focus_api_v1_macro_focus_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Focus",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/focus/assign": {
      "post": {
        "operationId": "assign_focus_view_api_v1_macro_focus_assign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FocusAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FocusAssignmentView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Focus View",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/lineage": {
      "get": {
        "operationId": "lineage_api_v1_macro_lineage_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lineage",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/runs": {
      "get": {
        "operationId": "macro_runs_api_v1_macro_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "world_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "World Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MacroRunsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Macro Runs",
        "tags": [
          "macro"
        ]
      }
    },
    "/api/v1/macro/schedules/{schedule_id}/cancel": {
      "post": {
        "operationId": "cancel_schedule_api_v1_macro_schedules__schedule_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Schedule",
        "tags": [
          "macro"
        ]
      }
    },
...
  }
}